chore: rename client-repo.ts to repo.ts
