+++ src/index.ts
@@ -3,4 +3,4 @@ export * from './types';
 export * from './client';
 export * from './result';
 export * from './query';
-export * from './client-repo';
+export * from './repo';
